+++ app/src/main/java/com/example/tutorsnotebook/views/fragments/tutor/AddStudentFragment.kt
@@ -79,7 +79,7 @@ class AddStudentFragment : Fragment() {
                     studentPhone = rootView.findViewById<EditText>(R.id.student_number_field).text.toString()
                         .toLong(), //TODO: add:regex
                     parentPhone = rootView.findViewById<EditText>(R.id.parent_number_field).text.toString()
-                        .toInt()
+                        .toLong()
                 )
 
                 Database.writeNewStudent(student)
